2004/10/05 Morgan Christiansson <drftpd@mog.se> * Fixed purge with JSXUserManager. * Added wkly_allotment to SITE USER environment and rewrote .properties format   for SITE USER output. * Fixed time to complete output in SiteBot and reviewed all other elapsed outputs. * Changed default permission in SITE USERS permission handling.   Defaults to the inverse of the last permission.   Ex: SITE USERS =deleted has !* as default       SITE USERS !=deleted has * as default
2004/10/03 Morgan Christiansson <drftpd@mog.se>
* Fixed SiteBot to use section name instead of section path when announcing.
  Thanks to N3m3Sis.

+++ src/net/sf/drftpd/master/UploaderPosition.java
@@ -20,7 +20,7 @@ package net.sf.drftpd.master;
 
 /**
  * @author mog
- * @version $Id: UploaderPosition.java,v 1.6 2004/08/03 20:13:56 zubov Exp $
+ * @version $Id: UploaderPosition.java,v 1.7 2004/10/05 02:11:22 mog Exp $
  */
 public class UploaderPosition implements Comparable {
     long bytes;
@@ -81,7 +81,7 @@ public class UploaderPosition implements Comparable {
             return 0;
         }
 
-        return (long) ((long) getBytes() / ((long) getXfertime() / 1000.0));
+        return (long) (getBytes() / (getXfertime() / 1000.0));
     }
 
     public long getXfertime() {
